% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtask.R
\name{action2entropy}
\alias{action2entropy}
\title{Step 1 of Subtask Analysis: obtaining entropy sequences of action sequences}
\usage{
action2entropy(action_seqs, rnn_dim = 20, n_epoch = 50,
  step_size = 0.001, batch_size = 1, optimizer_name = "rmsprop",
  index_valid = 0.2, verbose = FALSE)
}
\arguments{
\item{action_seqs}{a list of action sequences}

\item{rnn_dim}{latent dimension of RNN}

\item{n_epoch}{the number of training epochs.}

\item{step_size}{the learning rate of optimizer.}

\item{batch_size}{the batch size used in training.}

\item{optimizer_name}{a character string specifying the optimizer to be used
for training. Availabel options are \code{"sgd"}, \code{"rmsprop"}, 
\code{"adadelta"}, and \code{"adam"}.}

\item{index_valid}{proportion of sequences used as the validation set or a vector 
of indices specifying the validation set.}

\item{verbose}{logical. If TRUE, training progress is printed.}
}
\value{
\code{action2entropy} returns a list containing
\item{entropy_seqs}{a list of entropy sequences. The length of each entropy 
  sequence is one less than that of the corresponding action sequence.}
\item{loss_history}{a \code{n_epoch} by 2 matrix. The two columns contain
  the loss at the end of each epoch on the training set and the validation 
  set, respectively.}
\item{rnn_dim}{the latent dimension of the recurrent neural network}
\item{model_fit}{a vector of class \code{"raw"}. It is the serialized version of 
    the trained keras model.}
\item{actions}{a vector of the actions in \code{action_seqs}.}
\item{max_len}{maximum length of the action sequences.}
}
\description{
\code{action2entropy} fit a recurrent-neural-network-based action prediction 
model to a set of action sequences \code{action_seqs}.
}
\references{
Wang, Z., Tang, X., Liu, J., and Ying, Z. (2020) Subtask analysis of process data
  through a predictive model. \link{https://arxiv.org/abs/2009.00717}
}
\seealso{
\code{\link{entropy2segment}} and \code{\link{segment2subtask}} for 
  steps 2 and 3 of the subtask analysis procedure; \code{\link{subtask_analysis}} 
  for the complete procedure.
}
