% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_dist_cpp}
\alias{calculate_dist_cpp}
\title{calculate dissimilarity matrix}
\usage{
calculate_dist_cpp(seqs)
}
\arguments{
\item{seqs}{a list of action sequences}
}
\description{
calculate dissimilarity matrix
}
