% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_data}
\alias{cc_data}
\title{Data of item CP025Q01 (climate control item 1) in PISA 2012}
\format{A list with two elements. 
\describe{
  \item{seqs}{An object of class \code{"\link{proc}"} containing the action sequences and 
    the time sequences of the respondents.}
  \item{responses}{Binary responses of 16763 respondents. The order of the respondents
    matches that in \code{seqs}.}
}}
\source{
item interface: \url{http://www.oecd.org/pisa/test-2012/testquestions/question3/}
}
\usage{
cc_data
}
\description{
A dataset containing the response processes and binary response outcomes of 16763 respondents.
}
\keyword{datasets}
