% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{chooseK_mds}
\alias{chooseK_mds}
\title{Choose the number of MDS features by cross validation}
\usage{
chooseK_mds(seqs = NULL, K_cand, method = "oss", n_fold = 5,
  max_epoch = 100, step_size = 0.01, tot = 1e-06,
  return_dist = FALSE)
}
\arguments{
\item{seqs}{a list of \code{n} action sequences or their dissimilarity matrix.}

\item{K_cand}{candidates of number of features.}

\item{method}{the dissimilarity measure to be used to calculate distance matrix. Currently available methods: \code{oss}. It implements the order-based sequence similarity in Gomez-Alonso and Valls (2008).}

\item{n_fold}{number of folds for cross-validation}

\item{max_epoch}{the maximum number of epochs in stochastic gradient descent.}

\item{step_size}{the step size to be used in stochastic gradient descent}

\item{tot}{the tolerance of accuracy}

\item{return_dist}{logical. If dissimilarity matrix should be returned. Default is \code{FALSE}.}
}
\value{
a list
  \item{K}{number of features with smallest cross-validation loss}
  \item{K_cand}{candidates of number of features}
  \item{cv_loss}{cross-validation loss for each candidate K}
  \item{dist_mat}{dissimilarity matrix of sequences}
}
\description{
Choose the number of MDS features by cross validation
}
\examples{
n <- 50
seqs <- seq_gen(n)
K_res <- chooseK_mds(seqs, 5:10, return_dist=TRUE)
theta <- seq2feature_mds(K_res$dist_mat, K_res$K)$theta

}
