% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_functions.R
\name{count_actions}
\alias{count_actions}
\title{Action counts in a sequence}
\usage{
count_actions(x, actions)
}
\arguments{
\item{x}{an action sequences}

\item{actions}{a set of actions whose number of appearances will be count}
}
\description{
Action counts in a sequence
}
