% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{csv2seqs}
\alias{csv2seqs}
\title{Reading action sequences}
\usage{
csv2seqs(file, file_style, id_vars, seq_var, seq_sep = ",", ...)
}
\arguments{
\item{file}{the name of the file from which the action sequences are to be read.}

\item{file_style}{the style that the action sequences are stored. \code{"oneline"}
means each sequence occupies one line. Actions are separated by \code{seq_sep}.
\code{"multiline"} means that each sequence occupies multiple lines. Each line
represents an action in a sequence.}

\item{id_vars}{a string or a vector of strings giving the names of the variables
identify different sequences.}

\item{seq_var}{a string giving the name of the variable storing action sequences.}

\item{seq_sep}{the action separator characters. It is only used if \code{file_style="oneline"}.}

\item{...}{further arguments to be passed to \code{read.csv}.}
}
\value{
A list containing action sequences. Each element is an action sequence.
}
\description{
Reads a csv file and creates a list of action sequences.
}
\details{
This function first calls \code{read.csv} to read in data in the csv file and then
organize the action sequences stored in \code{seq_var} as a list of action sequences.
Each action sequence in the list is a vector of actions and is named by the concatenation
of the corresponsing values in \code{id_vars}.
}
