% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtask.R
\name{entropy2segment}
\alias{entropy2segment}
\title{Step 2 of Subtask Analysis: Segmenting Entropy Sequences}
\usage{
entropy2segment(entropy_seqs, lambda = 0.3, verbose = FALSE)
}
\arguments{
\item{entropy_seqs}{a list of entropy sequences}

\item{lambda}{a number between 0 and 1}

\item{verbose}{print progress if TRUE. default is FALSE}
}
\value{
a list containg the segment boundaries of each entropy sequence.
}
\description{
\code{entropy2segment} segments the entropy sequences in \code{entropy_seqs} 
using \code{segment_function}.
}
\references{
Wang, Z., Tang, X., Liu, J., and Ying, Z. (2020) Subtask analysis of process data
  through a predictive model. \link{https://arxiv.org/abs/2009.00717}
}
\seealso{
\code{\link{action2entropy}} and \code{\link{segment2subtask}} for 
  steps 1 and 3 of the subtask analysis procedure; \code{\link{subtask_analysis}} 
  for the complete procedure.
}
