% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_functions.R
\name{find_unigram}
\alias{find_unigram}
\title{Find all actions appears in a list of action sequences}
\usage{
find_unigram(seqs, count)
}
\arguments{
\item{seqs}{a list of action sequences}

\item{count}{logical value. If TRUE, unigram frequence will also be output}
}
\description{
Find all actions appears in a list of action sequences
}
