% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtask.R
\name{plot.subtask}
\alias{plot.subtask}
\title{Plot an subtask Object}
\usage{
\method{plot}{subtask}(object, type = "all", index = NULL,
  max_len = 5, col.subtask = 1:length(object$subtasks),
  cex.action = 0.5, lty = 1, pch = 16, srt = -90,
  plot_legend = TRUE, legend_pos = "topright", ...)
}
\arguments{
\item{object}{an object of class \code{"subtask"}}

\item{type}{\code{"all"} or \code{"individual"}}

\item{index}{a vector of indices of sequences to plot}

\item{max_len}{maximum length of plotted subtasks}

\item{col.subtask}{a vector of colors for subtasks}

\item{lty}{line types}

\item{pch}{point characters}

\item{plot_legend}{a logical value. If \code{TRUE} (default), plot the legend of subtasks.}

\item{legend_pos}{a character string or the coordinates to be used to position the legend.}

\item{...}{other arguments passed to \code{\link{legend}}}
}
\value{
this function does not return values
}
\description{
Plot the subtask analysis results for either the entire dataset or individual sequences.
}
\seealso{
\code{\link{plot_subtask_seq}}, \code{\link{plot_subtask_seqs}}.
}
