% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtask.R
\name{plot_subtask_seq}
\alias{plot_subtask_seq}
\title{Plot Subtask Analysis Results for One Sequence}
\usage{
plot_subtask_seq(action_seq, entropy_seq, subtask_seq, subtasks,
  col.subtask = 1:length(subtasks), cex.action = 0.5, lty = 1,
  pch = 16, srt = -90, plot_legend = TRUE, legend_pos = "topleft",
  ...)
}
\arguments{
\item{action_seq}{an action sequence}

\item{entropy_seq}{an entropy sequence}

\item{subtask_seq}{a subtask sequence}

\item{subtasks}{a vector of all subtasks}

\item{col.subtask}{a vector of colors for subtasks}

\item{lty}{line types}

\item{pch}{point characters}

\item{plot_legend}{a logical value. If \code{TRUE} (default), plot the legend of subtasks.}

\item{legend_pos}{a character string or the coordinates to be used to position the legend.}

\item{...}{other arguments passed to \code{\link{legend}}}
}
\value{
this function does not return values
}
\description{
Plot Subtask Analysis Results for One Sequence
}
\seealso{
\code{\link{plot_subtask_seqs}} for plotting results for all sequences. 
  \code{\link{plot.subtask}} for the plot method of \code{"subtask"} object.
}
