% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_object.R
\name{print.proc}
\alias{print.proc}
\title{Print method for class \code{"proc"}}
\usage{
\method{print}{proc}(x, n = 5, index = NULL, quote = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{"\link{proc}"}}

\item{n}{number of processes to be printed}
}
\description{
Print method for class \code{"proc"}
}
