% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_object.R
\name{print.summary.proc}
\alias{print.summary.proc}
\title{Print method for class \code{"summary.proc"}}
\usage{
\method{print}{summary.proc}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"\link{proc}"}}

\item{...}{not used.}
}
\value{
No return value.
}
\description{
Print method for class \code{"summary.proc"}
}
