% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcData.R
\docType{package}
\name{ProcData-package}
\alias{ProcData}
\alias{ProcData-package}
\title{ProcData: A package for process data analysis}
\description{
General tools for exploratory process data analysis. Process data refers to
the data describing participants' problem solving processes in computer-based
assessments. It is often recorded in computer log files. This package
provides two action sequence generators and implements two automatic feature
extraction methods that compress the information stored in process data,
which often has a nonstandard format, into standard numerical vectors. This
package also provides recurrent neural network based models that relate
response processes with other binary or scale variables of interest. The
functions that involve training and evaluating neural networks are wrappers
of functions in keras.
}
\section{Sequence generator}{
 two generators
}

\section{Feature extraction methods}{
 two feature extraction methods
}

\section{Recurrent neural network based models}{
 binary and scale
}

\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/xytangtang/ProcData/issues}
}

}
\author{
\strong{Maintainer}: Xueying Tang \email{xueyingtang1989@gmail.com}

}
