% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{read.seqs}
\alias{read.seqs}
\title{Reading action sequences}
\usage{
read.seqs(file, style, id_vars = "ID", seq_var = "Event", seq_sep = ",",
  ...)
}
\arguments{
\item{file}{the name of the csv file from which the action sequences are to be read.}

\item{style}{the style that the action sequences are stored. \code{"single"}
means each sequence occupies a single line. Actions are separated by \code{seq_sep}.
\code{"multiple"} means that each sequence occupies multiple lines, each action in
a line.}

\item{id_vars}{a string or a vector of strings giving the names of the variables
identify different sequences.}

\item{seq_var}{a string giving the name of the variable storing action sequences.}

\item{seq_sep}{the action separator characters. It is only used if \code{style="single"}.}

\item{...}{further arguments to be passed to \code{read.csv}.}
}
\value{
A list containing action sequences. Each element is an action sequence.
}
\description{
Reads a csv file and creates a list of action sequences.
}
\details{
This function first calls \code{read.csv} to read in data in the csv file and then
organize the action sequences stored in \code{seq_var} as a list of action sequences.
Each action sequence in the list is a vector of actions and is named by the concatenation
of the corresponsing values in \code{id_vars}.
}
