% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_manipulation.R
\name{remove_repeat}
\alias{remove_repeat}
\title{Remove repeated actions}
\usage{
remove_repeat(seqs, ignore = NULL)
}
\arguments{
\item{seqs}{an object of class \code{"\link{proc}"}}

\item{ignore}{repeated actions in ignore will not be deleted.}
}
\value{
an object of class \code{"\link{proc}"}
}
\description{
Remove repeated actions
}
