% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_manipulation.R
\name{replace_action}
\alias{replace_action}
\title{Replace actions in response processes}
\usage{
replace_action(seqs, old_action, new_action)
}
\arguments{
\item{seqs}{an object of class \code{"\link{proc}"}}

\item{old_action}{a string giving the action to be replaced.}

\item{new_action}{a string giving the action replacing \code{old_action}}
}
\value{
an object of class \code{"\link{proc}"}
}
\description{
Replace \code{old_action} with \code{new_action} in \code{seqs}. Timestamp
sequences are not affected.
}
\examples{
seqs <- seq_gen(10)
new_seqs <- replace_action(seqs, "Start", "Begin")
}
