% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtask.R
\name{segment2subtask}
\alias{segment2subtask}
\title{Step 3 of Subtask Analysis: Grouping Segments}
\usage{
segment2subtask(action_seqs, seg_seqs, n_subtask, actions,
  verbose = FALSE, ...)
}
\arguments{
\item{action_seqs}{a list of action sequences}

\item{seg_seqs}{a list of segment locations}

\item{n_subtask}{the desired number of subtasks or a vector of candidate number of subtasks}

\item{actions}{a set of actions}

\item{verbose}{logical. If TRUE, training progress is printed.}

\item{...}{additional arguments passed to \code{kmeans}}
}
\value{
a list containing 
\item{n_subtask}{the number of subtasks}
\item{subtasks}{a vector of subtasks}
\item{subtask_seqs}{a list of subtask sequences}
\item{tot.withinss}{a vector of total within cluster sum of squares}
\item{relative_cluster_profiles}{a \code{n_subtask} by \code{length(actions)} matrix. Each
  row gives the action frequency profile of each subtask relative to the overall action 
  frequency profile}
}
\description{
\code{segment2subtask} clustering action sequence segments according to their action
frequency profiles. Each cluster forms a subtask.
}
\references{
Wang, Z., Tang, X., Liu, J., and Ying, Z. (2020) Subtask analysis of process data
  through a predictive model. \link{https://arxiv.org/abs/2009.00717}
}
\seealso{
\code{\link{action2entropy}} and \code{\link{segment2subtask}} for 
  steps 1 and 3 of the subtask analysis procedure; \code{\link{subtask_analysis}} 
  for the complete procedure.
}
