% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{seq2feature_imds}
\alias{seq2feature_imds}
\title{Feature Extraction by iterative mds (for large dataset)}
\usage{
seq2feature_imds(seqs, K, dist_type, m)
}
\description{
Feature Extraction by iterative mds (for large dataset)
}
