% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{seq2feature_mds}
\alias{seq2feature_mds}
\title{Feature extraction via multidimensional scaling}
\usage{
seq2feature_mds(seqs = NULL, K = 2, method = "oss_action",
  max_epoch = 100, step_size = 0.01, pca = TRUE, tot = 1e-06,
  return_dist = FALSE, seed = 12345)
}
\arguments{
\item{seqs}{a \code{"\link{proc}"} object or a square matrix. If a squared matrix is 
provided, it is treated as the dissimilary matrix of a group of response processes.}

\item{K}{the number of features to be extracted.}

\item{method}{a character string specifies the dissimilarity measure for two
response processes. See 'Details'.}

\item{max_epoch}{the maximum number of epochs for stochastic gradient
descent.}

\item{step_size}{the step size of stochastic gradient descent.}

\item{pca}{a logical scalar. If \code{TRUE}, the principal components of the
extracted features are returned.}

\item{tot}{the accuracy tolerance for determining convergence.}

\item{return_dist}{logical. If \code{TRUE}, the dissimilarity matrix will be
returned. Default is \code{FALSE}.}

\item{seed}{random seed.}
}
\value{
\code{seq2feature_mds} returns a list containing 
  \item{theta}{a numeric matrix giving the \code{K} extracted features or principal
  features. Each column is a feature.} 
  \item{loss}{the value of the multidimensional scaling objective function.}
  \item{dist_mat}{the dissimilary matrix. This element exists only if \code{return_dist=TRUE}.}
}
\description{
\code{seq2feature_mds} extracts \code{K} features from response processes by
multidimensional scaling.
}
\details{
If \code{method="oss_action"}, order-based sequence similarity (oss) in 
Gomez-Alonso and Valls (2008) is used for action sequences. If
\code{method="oss_both"}, both action sequences and timestamp sequences are
used to a time-weighted oss. 
  
This function minimizes the objective function by stochastic gradient
descent. The coordinates of the objects are extracted features. The number of
features to be extracted \code{K} can be selected by cross-validation using
\code{\link{chooseK_mds}}.
}
\examples{
n <- 50
seqs <- seq_gen(n)
theta <- seq2feature_mds(seqs, 5)$theta
}
\references{
Gomez-Alonso, C. and Valls, A. (2008). A similarity measure for sequences of
  categorical data based on the ordering of common elements. In V. Torra & Y. Narukawa (Eds.) 
  \emph{Modeling Decisions for Artificial Intelligence}, (pp. 134-145). Springer Berlin Heidelberg.
}
\seealso{
\code{\link{chooseK_mds}} for choosing \code{K}.

Other feature extraction methods: \code{\link{aseq2feature_seq2seq}},
  \code{\link{atseq2feature_seq2seq}},
  \code{\link{seq2feature_seq2seq}},
  \code{\link{tseq2feature_seq2seq}}
}
\concept{feature extraction methods}
