% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{seq2feature_mds_large}
\alias{seq2feature_mds_large}
\title{Feature Extraction by MDS for Large Dataset}
\usage{
seq2feature_mds_large(seqs, K, dist_type = "oss_action", subset_size,
  subset_method = "random", n_cand = 10, pca = TRUE, seed = 12345)
}
\arguments{
\item{seqs}{an object of class \code{"\link{proc}"}}

\item{K}{the number of features to be extracted.}

\item{dist_type}{a character string specifies the dissimilarity measure for two
response processes. See 'Details'.}

\item{subset_size}{two parameters used in the large data algorithm. See 'Details'
and \code{\link{seq2feature_mds_large}}.}

\item{subset_method}{a character string specifying the method for choosing the subset 
in the large data algorithm. See 'Details' and \code{\link{seq2feature_mds_large}}.}

\item{n_cand}{two parameters used in the large data algorithm. See 'Details'
and \code{\link{seq2feature_mds_large}}.}

\item{pca}{logical. If \code{TRUE} (default), the principal components of the
extracted features are returned.}

\item{seed}{random seed.}
}
\description{
Feature Extraction by MDS for Large Dataset
}
\seealso{
Other feature extraction methods: \code{\link{aseq2feature_seq2seq}},
  \code{\link{atseq2feature_seq2seq}},
  \code{\link{seq2feature_mds}},
  \code{\link{seq2feature_seq2seq}},
  \code{\link{tseq2feature_seq2seq}}
}
\concept{feature extraction methods}
