% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_functions.R
\name{seq_gen}
\alias{seq_gen}
\title{Sequence generator}
\usage{
seq_gen(n, action_set1 = c("OPT1_1", "OPT1_2", "OPT1_3"),
  action_set2 = c("OPT2_1", "OPT2_2"), answer_set = c("CHECK_A",
  "CHECK_B", "CHECK_C", "CHECK_D"), p1 = rep(1, length(action_set1)),
  p2 = rep(1, length(action_set2)), p_answer = rep(1,
  length(answer_set)), p_continue = 0.5, p_choose = 0.5)
}
\arguments{
\item{n}{number of sequences to be generated}

\item{action_set1}{choices for the first condition}

\item{action_set2}{choices for the second condition}

\item{answer_set}{choices for the answer}

\item{p1}{probability weights to sample from \code{action_set1}}

\item{p2}{probability weights to sample from \code{action_set2}}

\item{p_answer}{probability weights to sample from \code{answer_set}}

\item{p_continue}{probability of running another simulation}

\item{p_choose}{probability of choose an answer}
}
\value{
a list of \code{n} elements. Each element is a sequence of actions.
}
\description{
\code{seq_gen} generates action sequences similar to PISA data: choose the levels of two conditions to run simulations and select an answer
}
