% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_functions.R
\name{seq_gen2}
\alias{seq_gen2}
\title{Generating action sequence according to probability transition matrix}
\usage{
seq_gen2(n, Pmat = NULL, events = letters, start_index = 1,
  end_index = length(events), max_len = 200)
}
\arguments{
\item{n}{number of sequences to be generated}

\item{Pmat}{\code{N} by \code{N} probability transition matrix; if \code{Pmat} not supplied, each action is randomly drawn from \code{events} (excluding start).}

\item{events}{a set of \code{N} possible actions including start and end}

\item{start_index}{index of start in \code{events}}

\item{end_index}{index of end in \code{events}}

\item{max_len}{maximum length of generated sequences}
}
\value{
a list of \code{n} elements. Each element is a sequence of actions.
}
\description{
Generating action sequence according to probability transition matrix
}
