% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{seq_predict}
\alias{seq_predict}
\title{Sequence model prediction}
\usage{
seq_predict(fit_res, new_seqs, batch_size = 16)
}
\arguments{
\item{fit_res}{An object returned by \code{seq2binary} or \code{seq2scale} with return_model=TRUE.}

\item{new_seqs}{Sequences that prediction is based on.}

\item{batch_size}{batch size for computing predicted values}
}
\value{
A vector of predicted values.
}
\description{
\code{seq_predict} predict response variables from a model fit by \code{seq2binary} or \code{seq2scale}.
}
\examples{
n <- 100
seqs <- seq_gen(n)
y1 <- sapply(seqs, function(x) "CHECK_A" \%in\% x)
y2 <- log10(sapply(seqs, length))
index_train <- 1:70
index_valie <- 71:85
index_test <- 86:100
res1 <- seq2binary(seqs, y1, index_train = index_train, index_valid = index_valid, index_test = index_valid)
res2 <- seq2scale(seqs, y2, index_train = index_train, index_valid = index_valid, index_test = index_valid)
res1_pred <- seq_predict(res1, seqs[index_test])
res2_pred <- seq_predict(res2, seqs[index_test])
all.equal(res1$pred, res1_pred) 
all.equal(res2$pred, res2_pred)

}
