% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_manipulation.R
\name{sub_seqs}
\alias{sub_seqs}
\title{Subset response processes}
\usage{
sub_seqs(seqs, ids)
}
\arguments{
\item{seqs}{an object of class \code{"\link{proc}"}}

\item{ids}{a vector of indices}
}
\value{
an object of class \code{"\link{proc}"}
}
\description{
Subset response processes
}
\examples{
data(cc_data)
seqs <- sub_seqs(cc_data$seqs, 1:10)
}
