% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtask.R
\name{subtask_analysis}
\alias{subtask_analysis}
\title{Subtask Analysis}
\usage{
subtask_analysis(action_seqs, lambda = 0.3, n_subtask, rnn_dim = 20,
  n_epoch = 20, step_size = 0.001, batch_size = 1,
  optimizer_name = "rmsprop", index_valid = 0.2, verbose = FALSE,
  ...)
}
\arguments{
\item{action_seqs}{a list of action sequences}

\item{lambda}{a number between 0 and 1}

\item{n_subtask}{the desired number of subtasks or a vector of candidate number of subtasks}

\item{rnn_dim}{latent dimension of RNN}

\item{n_epoch}{the number of training epochs.}

\item{step_size}{the learning rate of optimizer.}

\item{batch_size}{the batch size used in training.}

\item{optimizer_name}{a character string specifying the optimizer to be used
for training. Availabel options are \code{"sgd"}, \code{"rmsprop"}, 
\code{"adadelta"}, and \code{"adam"}.}

\item{index_valid}{proportion of sequences used as the validation set or a vector 
of indices specifying the validation set.}

\item{verbose}{logical. If TRUE, training progress is printed.}

\item{...}{additional arguments passed to \code{kmeans}}
}
\value{
an object of class \code{"subtask"}. It is a list containing
\item{action_seqs}{a list of action sequences}
\item{entropy_seqs}{a list of entropy sequences}
\item{seg_seqs}{a list of segment boundaries}
\item{subtask_seqs}{a list of subtask sequences}
\item{subtasks}{a vector of subtasks}
\item{n_subtask}{the number of subtasks}
\item{tot.withinss}{a vector of total within cluster sum of squares}
\item{relative_cluster_profiles}{a \code{n_subtask} by \code{length(actions)} matrix. Each
  row gives the action frequency profile of each subtask relative to the overall action 
  frequency profile}
\item{loss_history}{a \code{n_epoch} by 2 matrix. The two columns contain
  the loss at the end of each epoch on the training set and the validation 
  set, respectively.}
\item{rnn_dim}{the latent dimension of the recurrent neural network}
\item{model_fit}{a vector of class \code{"raw"}. It is the serialized version of 
    the trained action prediction model.}
\item{actions}{a vector of the actions in \code{action_seqs}.}
\item{max_len}{maximum length of the action sequences.}
}
\description{
\code{subtask_analysis} performs subtask identification procedure.
}
\references{
Wang, Z., Tang, X., Liu, J., and Ying, Z. (2020) Subtask analysis of process data
  through a predictive model. \link{https://arxiv.org/abs/2009.00717}
}
\seealso{
\code{\link{action2entropy}}, \code{\link{entropy2segment}}, and 
  \code{\link{segment2subtask}} for the three steps of subtask analysis.
}
