% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_functions.R
\name{time_seqs_summary}
\alias{time_seqs_summary}
\title{Summarize timestamp sequences}
\usage{
time_seqs_summary(time_seqs)
}
\arguments{
\item{time_seqs}{a list of timestamp sequences}
}
\value{
a list containing the following objects
  \item{total_time}{total time elapsed}
  \item{time_per_action}{average time between two consecutive actions}
  \item{time_interval_seqs}{inter-arrival time sequences}
}
\description{
Summarize timestamp sequences
}
