% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_functions.R
\name{tseq2interval}
\alias{tseq2interval}
\title{Transform a timestamp sequence to a time interval sequence}
\usage{
tseq2interval(x)
}
\arguments{
\item{x}{a timestamp sequence}
}
\description{
Transform a timestamp sequence to a time interval sequence
}
