% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{write.seqs}
\alias{write.seqs}
\title{Write process data to csv files}
\usage{
write.seqs(seqs, file, style, id_var = "ID", action_var = "Event",
  time_var = "Time", step_sep = ",", ...)
}
\arguments{
\item{seqs}{an object of class \code{"\link{proc}"} to written in the csv file.}

\item{file}{the name of the csv file from which the response processes are to be read.}

\item{style}{the style that the response processes are stored. See 'Details'.}

\item{id_var}{a string giving the name of the variable storing the process identifier.}

\item{action_var}{a string giving the name of the variable storing action sequences.}

\item{time_var}{a string giving the name of the variable storing timestamp sequences.}

\item{step_sep}{the step separator characters. It is only used if \code{style="single"}.}

\item{...}{further arguments to be passed to \code{write.csv}}
}
\value{
No return value.
}
\description{
Write process data to csv files
}
