% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_functions.R
\name{action_seqs_summary}
\alias{action_seqs_summary}
\title{Summarize action sequences}
\usage{
action_seqs_summary(action_seqs)
}
\arguments{
\item{action_seqs}{a list of action sequences}
}
\value{
a list containing the following objects:
  \item{n}{number of action sequences}
  \item{action}{action set}
  \item{seq_length}{sequence lengths}
  \item{count}{action counts}
  \item{count_by_seq}{action counts in each sequence}
  \item{seq_count}{counts of number of sequences with a given action}
  \item{seq_count_by_seq}{if actions appears in a sequence}
  \item{trans_count}{action transition counts}
}
\description{
Summarize action sequences
}
