% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_functions.R
\name{action_seqs_summary}
\alias{action_seqs_summary}
\title{Summarize action sequences}
\usage{
action_seqs_summary(action_seqs)
}
\arguments{
\item{action_seqs}{a list of action sequences.}
}
\value{
a list containing the following objects:
  \item{n_seq}{the number of action sequences}
  \item{n_action}{the number of distinct actions}
  \item{action}{the action set}
  \item{seq_length}{sequence lengths}
  \item{action_freq}{action counts}
  \item{action_seqfreq}{the number of sequences that each action appears}
  \item{trans_count}{a \code{length(action)} by \code{length(action)} matrix whose
    element in the i-th row and j-th column is the counts of transition from 
    \code{action[i]} to \code{action[j]}.}
}
\description{
Summarize action sequences
}
\seealso{
\code{\link{time_seqs_summary}} for summarizing timestamp sequences.
}
