% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_functions.R
\name{count_actions}
\alias{count_actions}
\title{count appearances of actions in \code{actions} in an action sequence}
\usage{
count_actions(x, actions)
}
\arguments{
\item{x}{an action sequence}

\item{actions}{a set of actions whose number of appearances will be count}
}
\description{
count appearances of actions in \code{actions} in an action sequence
}
