% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_functions.R
\name{count_actions}
\alias{count_actions}
\title{Count action appearances}
\usage{
count_actions(x, actions)
}
\arguments{
\item{x}{an action sequence.}

\item{actions}{a set of actions whose number of appearances will be count.}
}
\value{
an integer vector of counts.
}
\description{
This function counts the appearances of each action in \code{actions} in 
action sequence \code{x}.
}
