% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_model.R
\name{predict.seqm}
\alias{predict.seqm}
\title{Predict method for sequence models}
\usage{
\method{predict}{seqm}(object, new_seqs, new_data, ...)
}
\arguments{
\item{object}{a fitted object of class \code{"seqm"} from \code{seqm}.}

\item{new_seqs}{a list of action sequences with which to predict.}

\item{new_data}{a dataframe in which to look for variables with which to predict.}

\item{...}{further arguments to be passed to \code{predict.keras.engine.training.Model}.}
}
\value{
a vector of predictions. If \code{response_type="binary"}, predictions are
  probabilities of the response variable being 1.
}
\description{
Obtains predictions from a fitted sequence model object.
}
\details{
It unserialize \code{object$model_fit} to obtain a keras model of class 
\code{"keras.engin.training.Model"} and then calls \code{predict} for this class
to obtain predictions.
}
\seealso{
\code{\link{seqm}} for fitting sequence models.
}
