% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{seq2feature_imds}
\alias{seq2feature_imds}
\title{Feature Extraction by iterative mds (for large dataset)}
\usage{
seq2feature_imds(seqs, K, dist_type, m)
}
\arguments{
\item{seqs}{a \code{"\link{proc}"} object or a square matrix. If a squared matrix is 
provided, it is treated as the dissimilary matrix of a group of response processes.}

\item{K}{the number of features to be extracted.}

\item{dist_type}{type of dissimilarity}

\item{m}{subset size}
}
\description{
Feature Extraction by iterative mds (for large dataset)
}
