% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{seq2feature_imds}
\alias{seq2feature_imds}
\title{Feature Extraction by iterative mds (for large dataset)}
\usage{
seq2feature_imds(seqs, K, dist_type = "oss_action", m,
  subset_method = "random", n_sample, pca = TRUE)
}
\arguments{
\item{seqs}{a \code{"\link{proc}"} object or a square matrix. If a squared matrix is 
provided, it is treated as the dissimilary matrix of a group of response processes.}

\item{K}{the number of features to be extracted.}

\item{dist_type}{type of dissimilarity}

\item{m}{subset size}

\item{subset_method}{method for choosing subset size}

\item{n_sample}{candidate sample size}

\item{pca}{a logical scalar. If \code{TRUE}, the principal components of the
extracted features are returned.}
}
\description{
Feature Extraction by iterative mds (for large dataset)
}
