% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{seq2feature_ngram}
\alias{seq2feature_ngram}
\title{ngram feature extraction}
\usage{
seq2feature_ngram(seqs, level = 2, type = "binary", sep = "\\t")
}
\arguments{
\item{seqs}{an object of class \code{"\link{proc}"}}

\item{level}{an integer specifying the max length of ngrams}

\item{type}{a character string (\code{"binary"}, \code{"freq"}, or \code{"weighted"}) specifying the type of ngram features.}
}
\value{
a matrix of ngram features
}
\description{
ngram feature extraction
}
\examples{
seqs <- seq_gen(100)
theta <- seq2feature_ngram(seqs)
}
\seealso{
Other feature extraction methods: \code{\link{aseq2feature_seq2seq}},
  \code{\link{atseq2feature_seq2seq}},
  \code{\link{seq2feature_mds_large}},
  \code{\link{seq2feature_mds}},
  \code{\link{seq2feature_seq2seq}},
  \code{\link{tseq2feature_seq2seq}}
}
\concept{feature extraction methods}
