% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{seq2scale}
\alias{seq2scale}
\title{Predict continuous variable from action sequences}
\usage{
seq2scale(seqs, response, model_output = "seq2scale_model.h5",
  rnn_type = "lstm", n_hidden = 0, K = 20, K_hidden = NULL,
  n_epoch = 20, batch_size = 16, optimizer = "rmsprop", index_train,
  index_valid, index_test, gpu = TRUE)
}
\arguments{
\item{seqs}{a list of \code{n} action sequences. Each element is an action sequence in the form of a vector of actions.}

\item{response}{the continuous response variable}

\item{model_output}{a character string giving the name of the file to store keras model.}

\item{rnn_type}{the type of recurrent neural network to be used for modeling action sequences. \code{"lstm"} for long-short term memory. \code{"gru"} for gated recurrent unit.}

\item{n_hidden}{the number of hidden layers in the feed-forward neural network.}

\item{K}{the latent dimension in the recurrent neural network.}

\item{K_hidden}{a vector of length \code{n_hidden} specifying the number of nodes in each hidden layers.}

\item{n_epoch}{the number of epochs to be run in training.}

\item{batch_size}{batch size in training.}

\item{optimizer}{a character string specifies the optimizer to be used for training.}

\item{index_train}{a vector of indices specifying the training set}

\item{index_valid}{a vector of indices specifying the validation set}

\item{index_test}{a vector of indices specifying the test set}

\item{gpu}{logical. If TRUE, use gpu (if available) for training}
}
\value{
a list
  \item{summary}{a 3 by 2 matrix summarizing the prediction performance. Rows of the matrix indicate training set, validation set and test set. Columns of the matrix indicate mean squared error and R-square.}
  \item{trace}{a \code{n_epoch} by 2 matrix giving the trace of model training. The two columns gives the trace for the training set and the validation set, respectively.}
  \item{pred_train}{fitted values for the training set.}
  \item{pred_valid}{predicted values for the validation set.}
  \item{pred_test}{predicted values for the test set.}
}
\description{
Predict continuous variable from action sequences
}
\examples{
n <- 50
seqs <- seq_gen(n)
y <- log10(sapply(seqs, length))
index_train <- sample(1:n, round(0.8*n))
index_valid <- sample(setdiff(1:n, index_train), round(0.1*n))
index_test <- setdiff(1:n, c(index_train, index_valid))
res <- seq2scale(seqs, y, index_train = index_train, index_valid = index_valid, index_test = index_valid)
}
