% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_object.R
\name{summary.proc}
\alias{summary.proc}
\title{Summary method for class \code{"proc"}}
\usage{
\method{summary}{proc}(seqs)
}
\arguments{
\item{seqs}{an object of class \code{\link{"proc"}}}
}
\value{
a list
}
\description{
Summary method for class \code{"proc"}
}
\seealso{
\link{action_seqs_summary} and \link{time_seqs_summary}
}
