% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_object.R
\name{summary.proc}
\alias{summary.proc}
\title{Summary method for class \code{"proc"}}
\usage{
\method{summary}{proc}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"\link{proc}"}.}

\item{...}{not used.}
}
\value{
a list. Its components are the components returned by \link{action_seqs_summary} and
  \link{time_seqs_summary}.
}
\description{
The summary of a "proc" object combines the summary of the action sequences and the summary
of the timestamp sequences.
}
\seealso{
\link{action_seqs_summary} and \link{time_seqs_summary}
}
