% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_functions.R
\name{time_seqs_summary}
\alias{time_seqs_summary}
\title{Summarize timestamp sequences}
\usage{
time_seqs_summary(time_seqs)
}
\arguments{
\item{time_seqs}{a list of timestamp sequences}
}
\value{
a list containing the following objects
  \item{total_time}{a summary of response time of \code{n_seq} response processes}
  \item{mean_react_time}{a summary of mean reaction time of \code{n_seq} response processes}
}
\description{
Summarize timestamp sequences
}
